remove unused utils and merge them
