--- src/utils/index.ts
@@ -1,11 +1,0 @@
-export { round, floor, limit, limiter, multiply, percentage } from "./number";
-
-export { slugify, hexify } from "./string";
-
-export {
-  isClient,
-  isServer,
-  isValidHex,
-  isValidColor,
-  isValidCss,
-} from "./prefix";
